enterprise-scaffold/soc-api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
elasticsearch[async]==8.12.0
pydantic==2.8.2
pydantic-settings==2.5.2
